+++ backend/tatooine/modules/event/app/domain/ports/repositories/repositories.go
@@ -4,7 +4,7 @@ import "github.com/omerbeden/event-mate/backend/tatooine/modules/event/app/domai
 
 type EventRepository interface {
 	Create(event model.Event) (*model.Event, error)
-	GetByID(id int32) (*model.Event, error)
+	GetByID(id int64) (*model.Event, error)
 	GetByLocation(loc *model.Location) ([]model.Event, error)
 	UpdateByID(id int32, event model.Event) (bool, error)
 	DeleteByID(id int32) (bool, error)
